Add some explicit casts for some assignments between different integer types [patch from Oliver Eichler for compiling with Microsoft C].
git-svn-id: http://routino.org/svn/branches/MSVC@1669 cdd41072-61d7-4102-955e-fb406ba21da9

+++ src/files.c
@@ -616,7 +616,7 @@ int SkipFileBuffered(int fd,off_t skip)
 
  if((filebuffers[fd]->pointer+skip)>filebuffers[fd]->length)
    {
-    skip-=filebuffers[fd]->length-filebuffers[fd]->pointer;
+    skip-=(off_t)(filebuffers[fd]->length-filebuffers[fd]->pointer);
 
     filebuffers[fd]->pointer=0;
     filebuffers[fd]->length=0;
